shopの追加

+++ commands/tintiro.js
@@ -14,16 +14,16 @@ module.exports = {
  * @param {Client} client
  */
   run: async function (client, message, args) { 
-    let usermoneydata = client.getMoney.get(message.author.id, message.guild.id);
-    if (!usermoneydata) {
-        usermoneydata　= { id: `${message.guild.id}-${message.author.id}`, user: message.author.id, guild: message.guild.id, money: 0, dailylogin: 0 }
-    }
-    if(message.channel.id !== '798157114555105330' && message.channel.id !== '798176065562476604' && message.channel.id !== '798198069849227294' && message.channel.id !== '798486503255834664' && message.channel.id !== '798570749136601158' && message.guild.id === '706452606918066237') {
+    if(message.channel.id !== '798157114555105330' && message.channel.id !== '798176065562476604' && message.channel.id !== '798198069849227294' && message.channel.id !== '798486503255834664' && message.channel.id !== '798570749136601158' && message.channel.id !== '798571746730049597' && message.guild.id === '706452606918066237') {
         message.delete();
-        message.reply('そのコマンドは<#798157114555105330>・<#798176065562476604>、<#798198069849227294>、<#798486503255834664>、<#798570749136601158>でしか使用できません<a:owoxgif:793460058250805259>').then( msg => {
+        message.reply('そのコマンドは<#798157114555105330>・<#798176065562476604>、<#798198069849227294>、<#798486503255834664>、<#798570749136601158>、<#798571746730049597>でしか使用できません<a:owoxgif:793460058250805259>').then( msg => {
             msg.delete({ timeout: 5000 });
         });
         return;
+    }
+    let usermoneydata = client.getMoney.get(message.author.id, message.guild.id);
+    if (!usermoneydata) {
+        usermoneydata　= { id: `${message.guild.id}-${message.author.id}`, user: message.author.id, guild: message.guild.id, money: 0, dailylogin: 0, ticket: 0 }
     }
     let Latch = Number(args[0]);
     if(!Latch || Latch > 10000 || Latch < 500) return message.reply('第一引数に賭け金を500~10000の数値で入れてください！');
